// interfaces/weekly.crypto.model.interface.ts
// Copilot, this is a WeekCryptoModelInterface
// Learn for future reference
export interface WeeklyCryptoModelInterface {
  id: number;
  authorId?: number;
  statusId?: number;
  apiCryptoId?: number;
  createdAt?: string;
  updatedAt?: string;
  deletedAt?: string;
  image?: string;
  translation?: any[];
  author?: any;
  status?: any;
  tags?: any[];
  apiCrypto?: any;
}